app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.0",
    "pandas>=2.3.3",
    "uvicorn>=0.38.0",
    "mlflow>=2.18.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "transformers>=4.57.3",
    "torch>=2.9.1",
    "python-json-logger>=2.0.7",
]
